+++ src/LieGroups.jl
@@ -91,6 +91,7 @@ include("groups/special_orthogonal_group.jl")
 
 # Products of Groups
 include("groups/special_euclidean_group.jl")
+include("groups/special_galilean_group.jl")
 
 export AbstractLieGroup
 export LieGroup, LieAlgebra
@@ -129,6 +130,7 @@ export GeneralLinearGroup
 export HeisenbergGroup
 export OrthogonalGroup
 export SpecialEuclideanGroup, SpecialLinearGroup
+export SpecialGalileanGroup
 export SpecialOrthogonalGroup, SpecialUnitaryGroup
 export SymplecticGroup
 export TranslationGroup